Add branchToEnvironmentMapping to Jenkinsfiles
Now that quickstarters point to the production branch of the shared
library, they need to work with this version, not with the old tag.

Removing verbose is in line with this - the new version of the shared
library does not have a verbose setting anymore. It displays quite a
lot by default, and allows to enable debug mode if needed.

Fixes #69.