Import images after cloning instead of triggering builds
Now that builds are binary, we cannot easily trigger them. Besides, it
is much faster and safer to just import the images into the cloned
project.

According to
https://docs.openshift.com/container-platform/3.6/dev_guide/managing_images.html#importing-images-across-projects,
the best way to do this is to use `oc tag`.

Next to this change, I did a lot of cleanup - e.g. renamed some options
etc.