Configure PROTECTED_BRANCHES in template to make it more explicit
This way it is easier for users to see which branches are protected,
and also clearer how to change it. Also, it paves the way to allow
users to choose their own default in a future version of ODS when
those templates are applied via Tailor.

Related to https://github.com/opendevstack/ods-core/issues/107.