Update memory requests/limits
The default ratio was 10, which is quite high. This will lead OpenShift to oversubscribe the nodes to 10x their capacity. This can cause unexpected OOMKill events on customer applications, and very large high load induced instability in the entire cluster. Therefore, it is better to reduce the ratio (here to 1.5x).

I would love to reduce the default limit, but some quickstarters consume quite a lot of memory, so we need a high default. Maybe in the future, we can allow every quickstarter to choose this value individually (e.g. the frontend components running only Nginx need far less).